break up main
